Refactor openai responses
  Key Changes

  1. Moved files from openai/ into openai/responses/ subdirectory
  2. Extracted handle_non_streaming_response() from router.rs to responses/non_streaming.rs
  3. Created common.rs with shared SSE helpers used by both streaming and accumulator
  4. Removed unused re-exports of Provider, ProviderError, ProviderRegistry
  5. Updated documentation in openai_mcp_flow.md
