+++ sgl-model-gateway/src/routers/openai/mod.rs
@@ -7,15 +7,9 @@
 //! - Multi-turn tool execution loops
 //! - SSE (Server-Sent Events) streaming
 
-mod accumulator;
 mod context;
-pub mod mcp;
-pub mod provider;
-mod responses;
+mod provider;
+pub mod responses;
 mod router;
-mod streaming;
-mod tool_handler;
 
-// Re-export the main types for external use
-pub use provider::{Provider, ProviderError, ProviderRegistry};
 pub use router::OpenAIRouter;
