+++ sgl-model-gateway/src/routers/openai/responses/streaming.rs
@@ -21,16 +21,15 @@ use tokio::sync::mpsc;
 use tokio_stream::wrappers::UnboundedReceiverStream;
 use tracing::warn;
 
-// Import from sibling modules
-use super::accumulator::StreamingResponseAccumulator;
 use super::{
-    context::{RequestContext, StreamingEventContext, StreamingRequest},
+    accumulator::StreamingResponseAccumulator,
+    common::{extract_output_index, get_event_type, parse_sse_block, ChunkProcessor},
     mcp::{
         build_resume_payload, execute_streaming_tool_calls, inject_mcp_metadata_streaming,
         prepare_mcp_payload_for_streaming, send_mcp_list_tools_events, ToolLoopState,
     },
-    responses::{mask_tools_as_mcp, patch_streaming_response_json, rewrite_streaming_block},
     tool_handler::{StreamAction, StreamingToolHandler},
+    utils::{mask_tools_as_mcp, patch_streaming_response_json, rewrite_streaming_block},
 };
 use crate::{
     protocols::{
@@ -43,115 +42,10 @@ use crate::{
     routers::{
         header_utils::{apply_request_headers, preserve_response_headers},
         mcp_utils::{ensure_request_mcp_client, McpLoopConfig},
+        openai::context::{RequestContext, StreamingEventContext, StreamingRequest},
         persistence_utils::persist_conversation_items,
     },
 };
-
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
-/// Extract output_index from a JSON value
-#[inline]
-pub(super) fn extract_output_index(value: &Value) -> Option<usize> {
-    value.get("output_index")?.as_u64().map(|v| v as usize)
-}
-
-/// Get event type from event name or parsed JSON, returning a reference to avoid allocation
-#[inline]
-pub(super) fn get_event_type<'a>(event_name: Option<&'a str>, parsed: &'a Value) -> &'a str {
-    event_name
-        .or_else(|| parsed.get("type").and_then(|v| v.as_str()))
-        .unwrap_or("")
-}
-
-// ============================================================================
-// Chunk Processor
-// ============================================================================
-
-/// Processes incoming byte chunks into complete SSE blocks.
-/// Handles buffering of partial chunks and CRLF normalization.
-pub(super) struct ChunkProcessor {
-    pending: String,
-}
-
-impl ChunkProcessor {
-    pub fn new() -> Self {
-        Self {
-            pending: String::new(),
-        }
-    }
-
-    /// Append a chunk to the buffer, normalizing line endings
-    pub fn push_chunk(&mut self, chunk: &[u8]) {
-        let chunk_str = match std::str::from_utf8(chunk) {
-            Ok(s) => Cow::Borrowed(s),
-            Err(_) => Cow::Owned(String::from_utf8_lossy(chunk).into_owned()),
-        };
-        // Normalize CRLF to LF without extra allocation
-        let mut chars = chunk_str.chars().peekable();
-        while let Some(c) = chars.next() {
-            if c == '\r' && chars.peek() == Some(&'\n') {
-                // Skip \r when followed by \n
-                continue;
-            }
-            self.pending.push(c);
-        }
-    }
-
-    /// Extract the next complete SSE block from the buffer, if available
-    pub fn next_block(&mut self) -> Option<String> {
-        loop {
-            let pos = self.pending.find("\n\n")?;
-            let block = self.pending[..pos].to_string();
-            self.pending.drain(..pos + 2);
-
-            if !block.trim().is_empty() {
-                return Some(block);
-            }
-            // If block is empty, loop again to find the next one
-        }
-    }
-
-    /// Check if there's remaining content in the buffer
-    pub fn has_remaining(&self) -> bool {
-        !self.pending.trim().is_empty()
-    }
-
-    /// Take any remaining content from the buffer
-    pub fn take_remaining(&mut self) -> String {
-        std::mem::take(&mut self.pending)
-    }
-}
-
-// ============================================================================
-// SSE Parsing
-// ============================================================================
-
-/// Parse an SSE block into event name and data
-///
-/// Returns borrowed strings when possible to avoid allocations in hot paths.
-/// Only allocates when multiple data lines need to be joined.
-pub(super) fn parse_sse_block(block: &str) -> (Option<&str>, Cow<'_, str>) {
-    let mut event_name: Option<&str> = None;
-    let mut data_lines: Vec<&str> = Vec::new();
-
-    for line in block.lines() {
-        if let Some(rest) = line.strip_prefix("event:") {
-            event_name = Some(rest.trim());
-        } else if let Some(rest) = line.strip_prefix("data:") {
-            data_lines.push(rest.trim_start());
-        }
-    }
-
-    let data = if data_lines.len() == 1 {
-        Cow::Borrowed(data_lines[0])
-    } else {
-        Cow::Owned(data_lines.join("\n"))
-    };
-
-    (event_name, data)
-}
 
 // ============================================================================
 // Event Transformation and Forwarding
@@ -160,7 +54,7 @@ pub(super) fn parse_sse_block(block: &str) -> (Option<&str>, Cow<'_, str>) {
 /// Apply all transformations to event data in-place (rewrite + transform)
 /// Optimized to parse JSON only once instead of multiple times
 /// Returns true if any changes were made
-pub(super) fn apply_event_transformations_inplace(
+pub fn apply_event_transformations_inplace(
     parsed_data: &mut Value,
     ctx: &StreamingEventContext<'_>,
 ) -> bool {
@@ -390,7 +284,7 @@ fn send_buffered_arguments(
 
 /// Forward and transform a streaming event to the client
 /// Returns false if client disconnected
-pub(super) fn forward_streaming_event(
+pub fn forward_streaming_event(
     raw_block: &str,
     event_name: Option<&str>,
     data: &str,
@@ -525,7 +419,7 @@ fn maybe_inject_mcp_in_progress(
 
 /// Send final response.completed event to client
 /// Returns false if client disconnected
-pub(super) fn send_final_response_event(
+pub fn send_final_response_event(
     handler: &StreamingToolHandler,
     tx: &mpsc::UnboundedSender<Result<Bytes, io::Error>>,
     sequence_number: &mut u64,
@@ -582,7 +476,7 @@ pub(super) fn send_final_response_event(
 // ============================================================================
 
 /// Simple pass-through streaming without MCP interception
-pub(super) async fn handle_simple_streaming_passthrough(
+pub async fn handle_simple_streaming_passthrough(
     client: &reqwest::Client,
     circuit_breaker: &crate::core::CircuitBreaker,
     headers: Option<&HeaderMap>,
@@ -733,7 +627,7 @@ pub(super) async fn handle_simple_streaming_passthrough(
 }
 
 /// Handle streaming WITH MCP tool call interception and execution
-pub(super) async fn handle_streaming_with_tool_interception(
+pub async fn handle_streaming_with_tool_interception(
     client: &reqwest::Client,
     headers: Option<&HeaderMap>,
     req: StreamingRequest,
@@ -1075,7 +969,8 @@ pub(super) async fn handle_streaming_with_tool_interception(
     response
 }
 
-pub(super) async fn handle_streaming_response(ctx: RequestContext) -> Response {
+/// Main entry point for streaming responses
+pub async fn handle_streaming_response(ctx: RequestContext) -> Response {
     let worker = ctx.worker().expect("Worker not selected").clone();
     let circuit_breaker = worker.circuit_breaker();
     let headers = ctx.headers().cloned();
