+++ sgl-model-gateway/src/routers/openai/responses/tool_handler.rs
@@ -7,8 +7,8 @@ use tracing::warn;
 
 use super::{
     accumulator::StreamingResponseAccumulator,
+    common::{extract_output_index, get_event_type},
     mcp::FunctionCallInProgress,
-    streaming::{extract_output_index, get_event_type},
 };
 use crate::protocols::event_types::{
     is_function_call_type, FunctionCallEvent, OutputItemEvent, ResponseEvent,
@@ -20,7 +20,7 @@ use crate::protocols::event_types::{
 
 /// Action to take based on streaming event processing
 #[derive(Debug)]
-pub(crate) enum StreamAction {
+pub enum StreamAction {
     Forward,      // Pass event to client
     Buffer,       // Accumulate for tool execution
     ExecuteTools, // Function call complete, execute now
@@ -32,7 +32,7 @@ pub(crate) enum StreamAction {
 
 /// Maps upstream output indices to sequential downstream indices
 #[derive(Debug, Default)]
-pub(crate) struct OutputIndexMapper {
+pub struct OutputIndexMapper {
     next_index: usize,
     // Map upstream output_index -> remapped output_index
     assigned: HashMap<usize, usize>,
@@ -74,7 +74,7 @@ impl OutputIndexMapper {
 // ============================================================================
 
 /// Handles streaming responses with MCP tool call interception
-pub(super) struct StreamingToolHandler {
+pub struct StreamingToolHandler {
     /// Accumulator for response persistence
     pub accumulator: StreamingResponseAccumulator,
     /// Function calls being built from deltas
